Moved RecorderPanel MVC to recorder.py
